==> 1. Data Importing and Descriptive Analysis.ipynb <==
# 01. Importing libraries

# Import libraries
import pandas as pd
import numpy as np
import os

# 02. Importing data

# Create string from project folder path 
path = r'/Users/chadstacey/04-01-2021 InstaCart Basket Analysis'

# Create list 
vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']

# Import orders.csv using relevant shortcuts
df = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), usecols = vars_list)

# Print first 5 rows
df.head(5)

# Import products.csv
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

# Print first 20 rows of df_prods
df_prods.head(20)

# Print last 35 rows of df_prods
df_prods.tail(35)

# Print column names of df_prods
df_prods.columns

# Print number of rows and columns in df_prods
df_prods.shape

df_prods.describe()

df_prods.dtypes

11 a) The maximum value of the “aisle_id” column is 134

11 b) The data type of the “department_id” column is int64
