==> 2. Data Wrangling & Subsetting.ipynb <==
# 01. Importing libraries

# Import libraries
import pandas as pd
import numpy as np
import os

# 02. Importing data

# Create string from project folder path 
path = r'/Users/chadstacey/04-01-2021_InstaCart_Basket_Analysis'

# Create list 
vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']

# Import orders.csv
df_ords = pd.read_csv(os.path.join(path, '02_Data', 'Original_Data', 'orders.csv'), index_col = False)

# Import products.csv
df_prods = pd.read_csv(os.path.join(path, '02_Data', 'Original_Data', 'products.csv'), index_col = False)

# 03. Data wrangling examples from exercise

# Print functions within 'days_since_prior_order' column
df_ords['days_since_prior_order'].value_counts(dropna = False)

# Renaming column 'order_dow'
df_ords.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

df_ords.head()

# Remove 'eval_set' column
df_ords = df_ords.drop(columns = ['eval_set'])

df_ords.head()

# Change data type of 'order_id' column
df_ords['order_id'] = df_ords['order_id'].astype('str')

df_ords['order_id'].dtype

# Import departments.csv
df_dep = pd.read_csv(os.path.join(path, '02_Data', 'Original_Data', 'departments.csv'), index_col = False)

df_dep.head()

# Transposing df_dep
df_dep.T

# Altering header
df_dep_t = df_dep.T

df_dep_t

df_dep_t.reset_index()

new_header = df_dep_t.iloc[0]

df_dep_t_new = df_dep_t[1:]

df_dep_t_new

df_dep_t_new.columns = new_header

# Final step of transposing df_dep
df_dep_t_new

# Transform df_dep_t_new into a data dictionary 
data_dict = df_dep_t_new.to_dict('index')

data_dict

df_prods.head()

# Identify which department is signified by id of 19
print(data_dict.get('19'))

# List values that have a department_id of 19
df_snacks =  df_prods[df_prods['department_id']==19]

df_snacks.head()

# 04. Data wrangling questions from task

Question 2

# Change data type of 'user_id' column
df_ords['user_id'] = df_ords['user_id'].astype('str')

df_ords['user_id'].dtype

Question 3 (N.B. I think all the variables in df_ords have intuitive names but I will change one of them anyway just to show I can do it!)

# Change name of 'days_since_prior_order' without overwriting dataframe
df_ords.rename(columns = {'days_since_prior_order' : 'days_since_previous_order'})

Question 4 - As you can see below, the busiest hour for placing orders is hour 10, with 288,418 orders being placed during this hour

# Determine busiest hour for placing orders 
df_ords['order_hour_of_day'].value_counts(dropna = False)

Question 5 - a value of 4 in the "department_id" column within the df_prods represents 'produce'

# determine what value of '4' in the department_id column of df_prods represents
data_dict

Question 6

# Create subset containing breakfast item info 
df_breakfast =  df_prods[df_prods['department_id']==14]

df_breakfast.head()

Question 7

# Create subset containing alcohol, deli, beverages and meat/seafood item info
df_dinner_party =  df_prods[df_prods['department_id'].isin([5,7,12,20])]

df_dinner_party

Question 8 - As you can see above, there are 7,650 rows in the df_dinner_party dataframe

Question 9

# Create subset for user with user_id = 1
df_first_user = df_ords[df_ords['user_id']=='1']

df_first_user

Question 10

# Provide basic stats on first user's behaviour
df_first_user.describe()

Question 12

# Export df_ords dataframe as “orders_wrangled.csv” in “Prepared Data” folder
df_ords.to_csv(os.path.join(path, '02_Data','Prepared_Data', 'orders_wrangled.csv'))

Question 13

# Export df_dep_t_new dataframe as “departments_wrangled.csv” in “Prepared Data” folder
df_dep_t_new.to_csv(os.path.join(path, '02_Data','Prepared_Data', 'departments_wrangled.csv'))
