==> 3. Data Consistency Checks.ipynb <==
# 01. Importing libraries

# Import libraries
import pandas as pd
import numpy as np
import os

# 02. Importing data

# Create string from project folder path 
path = r'/Users/chadstacey/04-01-2021_InstaCart_Basket_Analysis'

# Create list 
vars_list = ['order_id', 'user_id', 'order_number', 'orders_day_of_week', 'order_hour_of_day', 'days_since_prior_order']

# Import products.csv
df_prods = pd.read_csv(os.path.join(path, '02_Data', 'Original_Data', 'products.csv'), index_col = False)

# Import orders.csv
df_ords = pd.read_csv(os.path.join(path, '02_Data', 'Prepared_Data', 'orders_wrangled.csv'), usecols = vars_list)

# 03. Exercise - Data Consistency Checks on df_prods

# Create a dataframe
df_test = pd.DataFrame()

# Create a mixed type column
df_test['mix'] = ['a', 'b', 1, True]

df_test.head()

# Check for mixed types
for col in df_test.columns.tolist():
  weird = (df_test[[col]].applymap(type) != df_test[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_test[weird]) > 0:
    print (col)

df_test['mix'] = df_test['mix'].astype('str')

df_test.head()

# Find missing values 
df_prods.isnull().sum()

# Create subset with null values
df_nan = df_prods[df_prods['product_name'].isnull()==True]

df_nan

df_prods.shape

# Create subset without null values
df_prods_clean = df_prods[df_prods['product_name'].isnull()==False]

df_prods_clean

# Look for duplicates
df_dups = df_prods_clean[df_prods_clean.duplicated()]

df_dups

df_prods_clean.shape

# Create subset without duplicates
df_prods_clean_no_dups = df_prods_clean.drop_duplicates()

df_prods_clean_no_dups.shape

# Export df_prods_clean_no_dups dataframe to “Prepared Data” folder
df_prods_clean_no_dups.to_csv(os.path.join(path, '02_Data','Prepared_Data', 'products_checked.csv'))

# 03. Task - Data Consistency Checks on df_ords

Question 2

df_prods.describe()

I don't think anything looks off in the above data. The count for all variables is 49,693. The minimum for all is 1 (as you'd probably expect). The 25%, 50% and 75% values make sense for all the variables. I know from a previous task that there are 21 department ID's so the max for this makes sense, and the maximum values for all the others look ok as well. Overall I'm happy with this data and I don't think anything needs to be investigated further.

EDIT: Having looked closer at the data, the max value and the standard deviation for price is off. How can the min value be 1, the mean value be 9.99 and the max be 99999? There is something wrong here and it needs to be investigated further. 

Question 3

# Check for mixed type data in df_ords
for col in df_ords.columns.tolist():
  weird = (df_ords[[col]].applymap(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords[weird]) > 0:
    print (col)

Question 4 - none of the columns have mixed type data so there is no need to fix it

Question 5

# Check for missing values in df_ords
df_ords.isnull().sum()

None of the columns have missing values apart from 'days_since_prior_order'. This would make sense - every time a purchase is made by a new customer, this column will be left blank. This column will only contain a value if the purchase has been made by a returning customer.  

Question 6

# Create new column (based on value in days_since_prior_order) stating whether it's a new customer or not 
df_ords['new_customer'] = np.where(df_ords.days_since_prior_order >= 0, False, True)

df_ords.head()

I used this method because the missing values are still important, they simply tell us whether the customer has purchased from InstaCart before. This new column states whether the purchase has been made by a new or returning customer. I used this url to help me with this function: https://cmdlinetips.com/2019/05/how-to-create-a-column-using-condition-on-another-column-in-pandas/

Alternatively I could create a subset with all the NaN values from days_since_prior_order:

# Create subset with all null values in days_since_prior_order column
df_dspo_nan = df_ords[df_ords['days_since_prior_order'].isnull()==True]

df_dspo_nan

# Check for duplicates in df_dspo_nan
df_dspo_nan_dups = df_dspo_nan[df_dspo_nan.duplicated()]

df_dspo_nan_dups

There are no duplicates, therefore all the missing values represent a new customer and no user_id's come up twice in this subset. 

Questions 7 and 8 

# Check for duplicates in df_ords
df_ords_dups = df_ords[df_ords.duplicated()]

df_ords_dups

There are no duplicate values therefore none that need to be addressed. 

Question 9 

del df_ords['new_customer']

df_ords

# Export cleaned df_ords to “Prepared Data” folder
df_ords.to_csv(os.path.join(path, '02_Data','Prepared_Data', 'orders_cleaned.csv'))
