==> 4. Combining and Exporting Data Part 1.ipynb <==
# 01. Importing libraries 

# Import libraries
import pandas as pd
import numpy as np
import os

# 02. Importing data 

# Create string from project folder path 
path = r'/Users/chadstacey/04-01-2021_InstaCart_Basket_Analysis'

# Create list 
vars_list = ['order_id', 'user_id', 'order_number', 'orders_day_of_week', 'order_hour_of_day', 'days_since_prior_order']

# Import orders.csv
df_ords = pd.read_csv(os.path.join(path, '02_Data', 'Prepared_Data', 'orders_cleaned.csv'), usecols = vars_list)

# Import df_ords_prior.csv
df_ords_prior = pd.read_csv(os.path.join(path, '02_Data', 'Original_Data', 'order_products__prior.csv'), index_col = False)

# 03. Checking output and shape 

# Check the output
df_ords.head()

# Check the output
df_ords_prior.head()

# Check shape
df_ords.shape

# Check shape
df_ords_prior.shape

# 04. Merging the data

# Merge df_ords and df_ords_prior
df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id')

# Check output
df_merged_large.head()

# Test without overwriting
pd.merge(df_ords_prior,df_ords, on = ['order_id'], indicator = True)

# Export data to pkl
df_merged_large.to_pickle(os.path.join(path, '02_Data','Prepared_Data', 'orders_products_combined.pkl'))
