==> 5. Combining and Exporting Data Part 2.ipynb <==
# 01. Importing libraries

# Import libraries
import pandas as pd
import numpy as np
import os

# 02. Importing data

# Create string from project folder path 
path = r'/Users/chadstacey/04-01-2021_InstaCart_Basket_Analysis'

# Import df_merged_large.pkl
df_merged_large = pd.read_pickle(os.path.join(path, '02_Data', 'Prepared_Data', 'orders_products_combined.pkl'))

# Check output
df_merged_large.head()

# Check shape
df_merged_large.shape

# Create list 
vars_list = ['product_id', 'product_name', 'aisle_id', 'department_id', 'prices']

# Import products_cleaned.csv
df_prods = pd.read_csv(os.path.join(path, '02_Data', 'Prepared_Data', 'products_checked.csv'), usecols = vars_list)

# Check output
df_prods.head()

# 03. Combine df_merged_large with df_prods

# Merge the two dataframes
df_orders_products_merged = df_merged_large.merge(df_prods, on = 'product_id')

# Check output
df_orders_products_merged.head()

df_orders_products_merged.shape

# Test merge without overwriting
pd.merge(df_orders_products_merged,df_prods, on = ['product_id'], indicator = True)

df_prods.head()

len(df_prods['product_id'])

df_prods.shape

df_orders_products_merged.shape

# Export data to pkl
df_orders_products_merged.to_pickle(os.path.join(path, '02_Data','Prepared_Data', 'orders_products_merged.pkl'))
