==> 6. Deriving New Variables.ipynb <==
# 01. Importing libraries

# Import libraries
import pandas as pd
import numpy as np
import os

# 02. Importing data

# Create string from project folder path 
path = r'/Users/chadstacey/04-01-2021_InstaCart_Basket_Analysis'

# Import orders_products_merged
df_ords_prods_merged = pd.read_pickle(os.path.join(path, '02_Data', 'Prepared_Data', 'orders_products_merged.pkl'))

df_ords_prods_merged.shape

# 03. Manipulating data

# Create subset of df_ords_prods_merged
df = df_ords_prods_merged[:1000000]

# Check output
df

# Create user-defined function for price range 
def price_label(row):
    if row['prices'] <= 5:
        return 'Low-range product'
    elif row['prices'] > 5 and row['prices'] <= 15:
        return 'Mid-range product'
    elif row ['prices'] > 15:
        return 'High-range product'
    else: return 'Not enough data'

# Apply the function 
df['price_range'] = df.apply(price_label, axis=1)

# Count how many of each price range 
df['price_range'].value_counts(dropna = False)

# Check for max value in prices 
df['prices'].max()

# Using loc() function for same purpose:

df.loc[df['prices'] > 15, 'price_range_loc'] = 'High-range product'

df.loc[(df['prices'] <= 15) & (df['prices'] > 5), 'price_range_loc'] = 'Mid-range product'

df.loc[df['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

# Count values
df['price_range_loc'].value_counts(dropna = False)

# Using loc() function on entire dataframe 
df_ords_prods_merged.loc[df_ords_prods_merged['prices'] > 15, 'price_range_loc'] = 'High-range product'

df_ords_prods_merged.loc[(df_ords_prods_merged['prices'] <= 15) & (df_ords_prods_merged['prices'] > 5), 'price_range_loc'] = 'Mid-range product'

df_ords_prods_merged.loc[df_ords_prods_merged['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

# Count values 
df_ords_prods_merged['price_range_loc'].value_counts(dropna = False)

# Print frequency of orders_day_of_the_week
df_ords_prods_merged['orders_day_of_week'].value_counts(dropna = False)

# Execute for-loop specifying busiest days 
result = []

for value in df_ords_prods_merged["orders_day_of_week"]:
  if value == 0:
    result.append("Busiest day")
  elif value == 4:
    result.append("Least busy")
  else:
    result.append("Regularly busy")

# Check output
result

# Create new column for busiest day 
df_ords_prods_merged['busiest_day'] = result

# Count values for busiest day 
df_ords_prods_merged['busiest_day'].value_counts(dropna = False)

# Check output
df_ords_prods_merged.head()

# 04. Questions from Task

Question 2

# Execute for-loop specifying busiest days 
result_two = []

for value in df_ords_prods_merged["orders_day_of_week"]:
  if value == 0 or value == 1:
    result_two.append("Busiest days")
  elif value == 3 or value == 4:
    result_two.append("Least busy days")
  else:
    result_two.append("Regularly busy")

# Check output
result_two

# Create new column for busiest days
df_ords_prods_merged['busiest_days'] = result_two

# Check output 
df_ords_prods_merged.head()

# Count values for busiest days
df_ords_prods_merged['busiest_days'].value_counts(dropna = False)

Question 3

The count above shows these columns are accurate. We know that the count for the least busy days (3,4) should be 3,787,193 + 3,844,096 which is equal to 7,631,289 and we know that the count for the most busy days (0,1) should be 5,665,830 + 6,209,632 which is equal to 11,875,462.

Question 4

# Print frequency of order_hour_of_day
df_ords_prods_merged['order_hour_of_day'].value_counts(dropna = False)

# Execute for-loop specifying busiest hours 
result_three = []

for value in df_ords_prods_merged["order_hour_of_day"]:
  if value == 10 or value == 11 or value == 14 or value == 15 or value == 13 or value == 12 or value == 16 or value == 9 == 17 or value == 8 or value == 18 or value == 19:
    result_three.append("Most orders")
  elif value == 20 or value == 7 or value == 21 or value == 22 or value == 23 or value == 6 or value == 0 or value == 1:
    result_three.append("Average orders")
  else:
    result_three.append("Fewest orders")

# Print result_three
result_three

# Create new column for busiest period of day
df_ords_prods_merged['busiest_period_of_day'] = result_three

# Check output
df_ords_prods_merged.head()

Question 5

df_ords_prods_merged['busiest_period_of_day'].value_counts(dropna = False)

Question 7

df_ords_prods_merged.shape

# Export data to pkl
df_ords_prods_merged.to_pickle(os.path.join(path, '02_Data','Prepared_Data', 'ords_prods_how_busy.pkl'))
