==> 7. Grouping Data & Aggregating Variables Part 1.ipynb <==
# 01. Importing libraries 

# Import libraries
import pandas as pd
import numpy as np
import os

# 02. Importing data

# Create string from project folder path 
path = r'/Users/chadstacey/04-01-2021_InstaCart_Basket_Analysis'

# Import orders_products_merged
ords_prods_merge = pd.read_pickle(os.path.join(path, '02_Data', 'Prepared_Data', 'ords_prods_how_busy.pkl'))

# Create subset 
df = ords_prods_merge[:1000000]

# Check shape
df.shape

# Check output
df.head()

# Group by product_name
df.groupby('product_name')

# Apply aggregate function to find mean
df.groupby('department_id').agg({'order_number': ['mean']})

# Use mean function instead
df.groupby('department_id')['order_number'].mean()

# Apply aggregate function to find mean, min, max
df.groupby('department_id').agg({'order_number': ['mean', 'min', 'max']})

# Create column with each user_id's max order
ords_prods_merge['max_order'] = ords_prods_merge.groupby(['user_id'])['order_number'].transform(np.max)

# Tell pandas not to assign any options regarding the maximum number of rows to display
pd.options.display.max_rows = None

# Check output
ords_prods_merge.head(100)

# Derive columns with loc()
ords_prods_merge.loc[ords_prods_merge['max_order'] > 40, 'loyalty_flag'] = 'Loyal customer'
ords_prods_merge.loc[(ords_prods_merge['max_order'] <= 40) & (ords_prods_merge['max_order'] > 10), 'loyalty_flag'] = 'Regular customer'
ords_prods_merge.loc[ords_prods_merge['max_order'] <= 10, 'loyalty_flag'] = 'New customer'

# Check output
ords_prods_merge[['user_id', 'loyalty_flag', 'max_order']].head(60)

ords_prods_merge.shape
