==> 10. More Data Visualisation with Python Part 1.ipynb <==
# 01. Importing libraries

import pandas as pd
import numpy as np
import os

# 02. Importing data

# Create path
path = r'/Users/chadstacey/04-01-2021_InstaCart_Basket_Analysis'

# Import customer dataset
df_cust = pd.read_csv(os.path.join(path, '02_Data', 'Original_Data', 'customers.csv'))

# Check output 
df_cust.head()

# Check shape 
df_cust.shape

# 03. Data wrangling

# Renaming column 'Surnam'
df_cust.rename(columns = {'Surnam' : 'Surname'}, inplace = True)

# Renaming column 'n_dependants'
df_cust.rename(columns = {'n_dependants' : 'dependants'}, inplace = True)

# Check output 
df_cust.head()

I've decided not to drop any columns as all could be useful in final analysis 

# Check data types
df_cust.dtypes

# 04. Data quality and consistency checks 

# Describe dataframe
df_cust.describe()

All these values look correct so I'm happy to move on to the next stage of cleaning. 

# check for mixed type data 
for col in df_cust.columns.tolist():
  weird = (df_cust[[col]].applymap(type) != df_cust[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_cust[weird]) > 0:
    print (col)

# Change dtype of First Name to string
df_cust['First Name'] = df_cust['First Name'].astype('str')

# Find missing values 
df_cust.isnull().sum()

No missing values :-)

# Look for duplicates
df_cust_dups = df_cust[df_cust.duplicated()]

df_cust_dups

No duplicates either :-)

# 05. Combine customer data with up-to-date InstaCart data

# Import up-to-date InstaCart data
ords_prods_merge = pd.read_pickle(os.path.join(path, '02_Data', 'Prepared_Data', 'ords_prods_with_flags.pkl'))

ords_prods_merge.head()

# Merge the two dataframes
ords_prods_cust = ords_prods_merge.merge(df_cust, on = 'user_id')

# Check shape 
ords_prods_cust.shape

# Check output 
ords_prods_cust.head()

ords_prods_cust.shape

# 06. Export merged dataframe

# Export data to pkl
ords_prods_cust.to_pickle(os.path.join(path, '02_Data','Prepared_Data', 'orders_products_customers_merged.pkl'))
